requirements.txt
# Core dependencies
python-dotenv==1.0.0

# Slack integration
slack-sdk==3.26.1

# Gmail integration
google-auth>=2.26.1,<3.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.111.0

# Timezone handling
pytz>=2022.1,<2023.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
hypothesis==6.92.1

# Mocking for tests
pytest-mock==3.12.0
